iso8583_message_manager/pyproject.toml
[project]
name = "iso8583_manager"
version = "0.1.0"
description = "A clean-architecture wrapper for pyiso8583"
authors = [
    { name = "Developer" }
]
requires-python = ">=3.10"
dependencies = [
    "iso8583-types",
    "iso8583-core",
    "pyiso8583>=1.0.1",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[tool.uv.sources]
iso8583-types = { path = "../packages/iso8583-types", editable = true }
iso8583-core = { path = "../packages/iso8583-core", editable = true }

[project.scripts]
iso8583-msg = "iso8583_manager.__main__:main"

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.0",
]
all = [
    "iso8583_manager[api,web]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "mypy>=1.0.0",
    "ruff>=0.3.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "openapi-spec-validator>=0.7",
    "schemathesis>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
